requirements.txt
Django==1.11
djangorestframework-jsonapi==2.3.1
djangorestframework
django-filter
django-cleanup==2.1.0
pyexcel-xlsx==0.5.3
pandas
numpy
sqlalchemy
sklearn
scipy==1.0.0